# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PostLapAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/google:
    post:
      operationId: authGoogle
      tags: [auth]
      summary: Authenticate with Google OAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleAuthInput"
      responses:
        "200":
          description: Auth success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Auth failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: User info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    GoogleAuthInput:
      type: object
      required: [credential]
      properties:
        credential:
          type: string
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, name, plan, trials_remaining, total_checks]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        plan:
          type: string
          enum: [free, pro]
        trials_remaining:
          type: integer
        total_checks:
          type: integer
        last_check_at:
          type: ["string", "null"]
...